(Name Mangling) Short name mangling is now hash-based instead of counter based
  * The option -rose:mangled:noshortname will stop ROSE from making L0R
    type short mangled names.  This could significantly slow down ROSE
    so use it only for debugging for now
  * Also removed a comment that was getting needlessly duplicated by
    ROSETTA in incorrect places
  * Added "make64Bits" to Combinatorics to make any hash return a uint64_t
  * Added toBase62String to convert uint64_t numbers to a base-62 number
    string for mangled names that are a short as possible but usable
    as C++ identifiers
  * Short mangled names are now hash based.
  * EasyStorage got new types to handle map<string, uint64_t>

RC-1585

+++ src/midend/astDiagnostics/AstConsistencyTests.C
@@ -4509,8 +4509,8 @@ TestMangledNames::visit ( SgNode* node )
      SgGlobal* global = isSgGlobal(node);
      if (global != NULL)
         {
-          const std::map<std::string, int>& shortMangledNameCache = global->get_shortMangledNameCache();
-          for (std::map<std::string, int>::const_iterator i = shortMangledNameCache.begin(); i != shortMangledNameCache.end(); ++i) {
+          const std::map<std::string, uint64_t>& shortMangledNameCache = global->get_shortMangledNameCache();
+          for (std::map<std::string, uint64_t>::const_iterator i = shortMangledNameCache.begin(); i != shortMangledNameCache.end(); ++i) {
             this->totalLongMangledNameSize += i->first.size();
             ++this->totalNumberOfLongMangledNames;
           }
@@ -4722,8 +4722,8 @@ TestMangledNames::test()
      TestMangledNames t;
 
   // DQ (6/26/2007): Added code by Jeremiah for shorter mangled names
-     const std::map<std::string, int>& shortMangledNameCache = SgNode::get_shortMangledNameCache();
-     for (std::map<std::string, int>::const_iterator i = shortMangledNameCache.begin(); i != shortMangledNameCache.end(); ++i)
+     const std::map<std::string, uint64_t>& shortMangledNameCache = SgNode::get_shortMangledNameCache();
+     for (std::map<std::string, uint64_t>::const_iterator i = shortMangledNameCache.begin(); i != shortMangledNameCache.end(); ++i)
         {
           t.totalLongMangledNameSize += i->first.size();
           ++(t.totalNumberOfLongMangledNames);
